Added sub-communicator runner to the Mesh test
