+++ tests/unit-tests/data/mpi/dim3/HomogeneousMediumMPI4x4/Database/proc_3.bin
