+++ tests/unit-tests/mesh/mpi/dim3/.gitignore
@@ -1,0 +1,1 @@
+!test_fixture.hpp
