+++ tests/unit-tests/mesh/mpi/dim3/runner.cpp
@@ -1,0 +1,14 @@
+#include "../../../SPECFEM_Environment.hpp"
+
+#include "test_fixture.hpp"
+#include <gtest/gtest.h>
+#include <string>
+
+INSTANTIATE_TEST_SUITE_P(MPIMesh3DTests, MPIMesh3DTest,
+                         ::testing::Values("HomogeneousMediumMPI4x4"));
+
+int main(int argc, char *argv[]) {
+  ::testing::InitGoogleTest(&argc, argv);
+  ::testing::AddGlobalTestEnvironment(new SPECFEMEnvironment(4));
+  return RUN_ALL_TESTS();
+}
